@@ -1,161 +1,126 @@
 [build-system]
-requires = ["setuptools>=64", "wheel", "build>=0.7.0"]
+requires = ["setuptools>=61.0", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
-# This section is defined by PEP 621
 name = "homebrew-versiontracker"
-dynamic = ["version"]
-description = "CLI versiontracker and update tool for macOS"
+version = "0.1.0-alpha.1"
+description = "A command-line tool for tracking and managing applications installed outside of the Mac App Store"
 readme = "README.md"
-requires-python = ">=3.10"
-license = "MIT"
-authors = [{ name = "docdyhr", email = "thomas@dyhr.com" }]
-keywords = ["homebrew", "versiontracker", "macos", "update", "cask", "mac-apps"]
+license = {text = "MIT"}
+requires-python = ">=3.12"
+authors = [
+    {name = "Your Name", email = "your.email@example.com"}
+]
+keywords = ["macos", "homebrew", "version", "tracking", "updates", "cask"]
 classifiers = [
-    "Development Status :: 4 - Beta",
+    "Development Status :: 3 - Alpha",
     "Environment :: Console",
-    "Intended Audience :: Developers",
     "Intended Audience :: System Administrators",
+    "Intended Audience :: End Users/Desktop",
+    "License :: OSI Approved :: MIT License",
     "Operating System :: MacOS :: MacOS X",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.10",
-    "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
-    "Programming Language :: Python :: 3.13",
+    "Topic :: System :: Software Distribution",
     "Topic :: System :: Systems Administration",
     "Topic :: Utilities",
 ]
+
 dependencies = [
-    "fuzzywuzzy>=0.18.0",
-    "rapidfuzz>=3.0.0",
-    "tqdm>=4.66.0",
-    "PyYAML>=6.0",
-    "termcolor>=2.3.0",
-    "tabulate>=0.9.0",
-    "psutil>=6.1.0,<8.0.0",
-    "aiohttp>=3.8.0",
+    "pyyaml>=6.0",
+    "click>=8.0",
+    "tqdm>=4.65",
+    "psutil>=5.9",
+]
+
+[project.optional-dependencies]
+dev = [
+    "pytest>=7.4",
+    "pytest-cov>=4.1",
+    "pytest-timeout>=2.1",
+    "pytest-mock>=3.11",
+    "black>=23.7",
+    "ruff>=0.1",
+    "mypy>=1.5",
+    "types-PyYAML>=6.0",
+    "build>=0.10",
+    "wheel>=0.41",
+    "twine>=4.0",
+]
+test = [
+    "pytest>=7.4",
+    "pytest-cov>=4.1",
+    "pytest-timeout>=2.1",
+    "pytest-mock>=3.11",
+]
+security = [
+    "bandit[toml]>=1.7",
+    "pip-audit>=2.6",
+    "safety>=2.3",
+]
+fuzzy = [
+    "rapidfuzz>=3.0",
 ]
 
 [project.urls]
-Homepage = "https://github.com/docdyhr/versiontracker"
-"Bug Tracker" = "https://github.com/docdyhr/versiontracker/issues"
-Documentation = "https://github.com/docdyhr/versiontracker"
-"Source Code" = "https://github.com/docdyhr/versiontracker"
-Changelog = "https://github.com/docdyhr/versiontracker/blob/main/CHANGELOG.md"
-PyPI = "https://pypi.org/project/macos-versiontracker/"
+Homepage = "https://github.com/yourusername/versiontracker"
+Documentation = "https://github.com/yourusername/versiontracker#readme"
+"Bug Reports" = "https://github.com/yourusername/versiontracker/issues"
+"Source Code" = "https://github.com/yourusername/versiontracker"
 
 [project.scripts]
-versiontracker = "versiontracker.__main__:main"
-macos-versiontracker = "versiontracker.__main__:main"
+versiontracker = "versiontracker.cli:main"
 
-[tool.setuptools.dynamic]
-version = {attr = "versiontracker.__version__"}
+[tool.setuptools]
+packages = ["versiontracker", "versiontracker.handlers"]
 
-[tool.setuptools.packages.find]
-where = ["."]
-include = ["versiontracker*"]
-exclude = [
-    "images*",
-    "tmp_test*",
-    "tests*",
-    "docs*",
-    "htmlcov*",
-    "tools*",
-    "handlers*",
-    "node_modules*",
-]
+[tool.setuptools.package-data]
+versiontracker = ["py.typed"]
+
+[tool.black]
+line-length = 88
+target-version = ["py312"]
+include = '\.pyi?$'
 
 [tool.ruff]
-# Define the target Python version.
-# We support Python 3.13 and later.
-target-version = "py310"
-
-# Line length configuration - harmonized for AI assistant best practices
-# 120 characters provides good readability while being AI-friendly
-line-length = 120
-
-# Exclude the venv directory.
-exclude = [
-    ".bzr",
-    ".direnv",
-    ".eggs",
-    ".git",
-    ".hg",
-    ".mypy_cache",
-    ".nox",
-    ".pants.d",
-    ".pytype",
-    ".ruff_cache",
-    ".svn",
-    ".tox",
-    ".venv",
-    "__pypackages__",
-    "_build",
-    "buck-out",
-    "build",
-    "dist",
-    "node_modules",
-    "venv",
-    "examples",
-]
+line-length = 120  # AI-friendly length as per CLAUDE.md guidelines
+target-version = "py310"  # Updated to match project requirements
 
 [tool.ruff.lint]
-select = ["E", "F", "I"]
-# E203: whitespace before ':' (handled by ruff formatter)
-# E402: module level import not at top (common in test files with conditional imports)
-# F401: unused import (keeping old config for iterative development)
-# F811: redefinition of unused name (keeping old config for iterative development)
-# F821: undefined name (keeping old config for iterative development)
-# F841: assigned but unused variable (keeping old config for iterative development)
-# Note: E501 (line too long) is enforced via line-length setting above
-ignore = ["E203", "E402", "F401", "F811", "F821", "F841"]
-
-# Allow autofix for all enabled rules (when --fix)
+select = [
+    "E",  # pycodestyle errors
+    "W",  # pycodestyle warnings
+    "F",  # pyflakes
+    "I",  # isort
+    "B",  # flake8-bugbear
+    "C4", # flake8-comprehensions
+    "UP", # pyupgrade
+    "C90", # mccabe complexity
+]
+ignore = [
+    "E501",  # Line too long (handled by line-length setting)
+]
 fixable = ["ALL"]
 unfixable = []
 
-# Allow unused variables when underscore-prefixed
-dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
+[tool.ruff.lint.mccabe]
+max-complexity = 10
 
-[tool.ruff.lint.isort]
-known-first-party = ["versiontracker"]
-
-[tool.ruff.format]
-# Use black formatting style
-quote-style = "double"
-indent-style = "space"
-skip-magic-trailing-comma = false
-line-ending = "auto"
-
-# MyPy configuration
 [tool.mypy]
-python_version = "3.10"                         # Use 3.10 as baseline for compatibility
-warn_return_any = false
+python_version = "3.12"
+warn_return_any = true
 warn_unused_configs = true
-disallow_untyped_defs = false
-disallow_incomplete_defs = false
-check_untyped_defs = false
-disallow_untyped_decorators = false
+disallow_untyped_defs = true
+disallow_any_unimported = false
 no_implicit_optional = true
-strict_optional = true
+check_untyped_defs = true
 warn_redundant_casts = true
-warn_unused_ignores = false
+warn_unused_ignores = true
 warn_no_return = true
-warn_unreachable = true
-strict_equality = true
-show_error_codes = true
-show_error_context = true
-color_output = true
-error_summary = true
-pretty = true
-follow_imports = "silent"
-namespace_packages = true
-explicit_package_bases = true
-ignore_missing_imports = true
-cache_dir = ".mypy_cache"
-show_column_numbers = true
-exclude = ["examples/", "versiontracker/apps/"]
+follow_imports = "normal"
+
+
 
 [[tool.mypy.overrides]]
 module = [
